fuzzer: add web interface to fuzzer

+++ utilities/fuzzer.py
@@ -37,11 +37,8 @@ class Config():
         if uname not in config.sections():
             uname = "DEFAULT"
 
-        self.random_tests = config[uname]['random_tests']
-
         # A list of clients-tuples: name , isDocker, path
         self.active_clients = []
-        logger.info("lcietn %s" % config[uname]['clients'])
         for c in config[uname]['clients'].split(","):
             key = "{}.binary".format(c)
             if key in config[uname]:
@@ -58,7 +55,6 @@ class Config():
             path = os.path.abspath(path)
             return path
 
-        self.logs_path = resolve(config[uname]['logs_path'])
         # Artefacts are where stuff is stored
         self.artefacts = resolve(config[uname]['artefacts'])
         # temp paths is where we put stuff we don't necessarily save
@@ -67,32 +63,35 @@ class Config():
         #here = os.path.dirname(os.path.realpath(__file__))
         self.host_id = "%s-%s-%d" % (uname, time.strftime("%a_%H_%M_%S"), os.getpid())
 
-        self.info()
+        logger.info("\n".join(self.info()))
 
-        os.makedirs(self.testfilesPath(), exist_ok = True)
-        os.makedirs(self.logs_path, exist_ok = True)
         os.makedirs(self.artefacts, exist_ok = True)
+        os.makedirs(self.testfilesPath(), exist_ok = True)
+        os.makedirs(self.logfilesPath(), exist_ok = True)
 
 
     def testfilesPath(self):
         return "%s/testfiles/" % self.temp_path
 
+    def logfilesPath(self):
+        return "%s/logs/" % self.temp_path
+
     def clientNames(self):
         return [name for (name,y,z) in self.active_clients]
 
     def info(self):
-        logger.info("Config")
-        logger.info("  Active clients:")
+        out = []
+        out.append("Active clients:")
         for (name, isDocker, path) in self.active_clients:
-            logger.info("\t* {} : {} docker:{}".format(name, path, isDocker) )
+            out.append("  * {} : {} docker:{}".format(name, path, isDocker) )
 
-        logger.info("  Test generator: native (py)")
-     
-        logger.info("  Fork config:   %s",self.fork_config)
-        logger.info("  Log path:      %s",self.logs_path)
-        logger.info("  Artefacts:     %s",self.artefacts)
-        logger.info("  Tempfiles:     %s",self.temp_path)
-        logger.info("  Test files:    %s",self.testfilesPath())
+        out.append("Test generator: native (py)")
+        out.append("Fork config:   %s" % self.fork_config)
+        out.append("Artefacts:     %s" % self.artefacts)
+        out.append("Tempfiles:     %s" % self.temp_path)
+        out.append("Log path:      %s" % self.logfilesPath())
+        out.append("Test files:    %s" % self.testfilesPath())
+        return out
 
 cfg =Config('statetests.ini')
 
@@ -109,6 +108,7 @@ class StateTest():
         self.canon_traces = []
         self.procs = []
         self.traceFiles = []
+        self.additionalArtefacts = []
 
     def id(self):
         return self.identifier
@@ -122,26 +122,28 @@ class StateTest():
 
     def writeToFile(self):
         # write to unique tmpfile
-        logger.info("Writing file %s" % self.fullfilename())
+        logger.debug("Writing file %s" % self.fullfilename())
         with open(self.fullfilename(), 'w') as outfile:
             json.dump(self.statetest, outfile)
 
     def removeFiles(self):
+        if True:
+            return
         f = self.fullfilename()
-        logger.info("Removing test file %s" % f)
+        logger.debug("Removing test file %s" % f)
         os.remove(f)
 
         #delete non-failed traces
         for f in self.traceFiles:
-            logger.info("Removing trace file %s" % f)
+            logger.debug("Removing trace file %s" % f)
             os.remove(f)
 
     def tempTraceLocation(self, client):
-        return os.path.abspath("%s/%s-%s.trace.log" % (cfg.logs_path,self.identifier, client))
+        return os.path.abspath("%s/%s-%s.trace.log" % (cfg.logfilesPath(),self.identifier, client))
 
-    def saveTrace(self, client, output, command):
+    def storeTrace(self, client, output, command):
         filename = self.tempTraceLocation(client)
-        logging.info("Writing %s full trace to %s" % (self.id(), filename ))
+        logging.debug("Writing %s full trace to %s" % (self.id(), filename ))
         with open(filename, "w+") as f: 
             f.write("# command\n")
             f.write("# %s\n\n" % command)
@@ -155,30 +157,31 @@ class StateTest():
         logger.info("Saving testcase as %s", saveloc)
         os.rename(self.fullfilename() , saveloc)
 
+        newTracefiles = []
+
         for f in self.traceFiles:
             fname = os.path.basename(f)
             newloc = "%s/%s" % (cfg.artefacts,fname)
             logger.info("Saving trace as %s", newloc)
             os.rename(f, newloc)
-
+            newTracefiles.append(newloc)
+    
+        self.traceFiles = newTracefiles
  
     def addArtefact(self, fname, data):
         fullpath = "%s/%s-%s" % (cfg.artefacts, self.id(), fname)
         logger.info("Saving artefact %s", fullpath)
         with open(fullpath, "w+") as f:
-            f.write(data)
+            f.write(data)    
+        self.additionalArtefacts.append(fullpath)
 
-
-def dumpJson(obj, dir = None, prefix = None):
-    import tempfile
-    fd, temp_path = tempfile.mkstemp(prefix = 'randomtest_', suffix=".json", dir = dir)
-    with open(temp_path, 'w') as f :
-        json.dump(obj,f)
-        logger.info("Saved file to %s" % temp_path)
-    os.close(fd)
-    return temp_path
-
-    
+    def listArtefacts(self):
+        return {
+            "id"   : self.id(),
+            "file" : self.filename(), 
+            "traces": [os.path.basename(f) for f in self.traceFiles],
+            "other" : [os.path.basename(f) for f in self.additionalArtefacts], 
+        }
 
 def generateTests():
     """This method produces json-files, each containing one statetest, with _one_ poststate. 
@@ -194,18 +197,13 @@ def generateTests():
     t = templates.new(templates.object_based.TEMPLATE_RandomStateTest)
     test = {}
     counter = 0
+
     while True: 
         test.update(t)
         test_obj = json.loads(json.dumps(t, cls=randomtest.RandomTestsJsonEncoder))
         s = StateTest(test_obj, counter)
         counter = counter +1
         yield s
-        #break
-
-def main():
-    # Start all docker daemons that we'll use during the execution
-    startDaemons()
-    perform_tests(generateTests)
 
 
 def get_summary(combined_trace, n=20):
@@ -226,8 +224,6 @@ def get_summary(combined_trace, n=20):
     return list(buf)
 
 def startDaemon(clientname, imagename):
-
-
     dockerclient.containers.run(image = imagename, 
         entrypoint="sleep",
         command = ["356d"],
@@ -286,11 +282,8 @@ def execInDocker(name, cmd, stdout = True, stderr=True):
     #logger.info("executing in %s: %s" %  (name," ".join(cmd)))
     container = dockerclient.containers.get(name)
     (exitcode, output) = container.exec_run(cmd, stream=stream,stdout=stdout, stderr = stderr)     
-    logger.info("Executing %s : done in %f seconds" % (name, time.time() - start_time))
-
-
+    logger.info("Executing %s took in %f seconds" % (name, time.time() - start_time))
     return {'output': [output], 'cmd':" ".join(cmd)}
-
 
 def startGeth(test):
     """
@@ -306,7 +299,7 @@ def startGeth(test):
     
 
 def startParity(test):
-    cmd = ["/parity-evm","state-test", "--std-json","/testfiles/%s" % os.path.basename(test.tmpfile)]
+    cmd = ["/parity-evm","state-test", "--std-json","/testfiles/%s" % os.path.basename(test.filename())]
     return execInDocker("parity", cmd)
 
 def startHera(test):
@@ -335,7 +328,7 @@ def start_processes(test):
 
     starters = {'geth': startGeth, 'cpp': startCpp, 'parity': startParity, 'hera': startHera}
 
-    logger.info("Starting processes for %s on test %s" % ( cfg.clientNames(), test.identifier))
+    logger.info("Starting processes for %s on test %s" % ( cfg.clientNames(), test.id()))
     #Start the processes
     for (client_name, x, y) in cfg.active_clients:
         if client_name in starters.keys():
@@ -365,27 +358,36 @@ def readOutput(processInfo):
 
 
 def end_processes(test):
+    """ End processes for the given test, slurp up the output and compare the traces
+    returns the length of the canon-trace emitted (or -1)
+    """
     # Handle the old processes
-    if test is not None:
-        for (proc_info, client_name) in test.procs:
-            canonicalizer = canonicalizers[client_name]
-            full_trace = readOutput(proc_info)
-            test.saveTrace(client_name, full_trace,proc_info['cmd'])
-            canon_trace = [VMUtils.toText(step) for step in canonicalizer(full_trace.split("\n"))]
-            test.canon_traces.append(canon_trace)
-            logger.info("Processed %s steps for %s on test %s" % (len(canon_trace), client_name, test.identifier))
-
+    if test is None: 
+        return None
+    tracelen = 0
+    for (proc_info, client_name) in test.procs:
+        logger.info("Proc '%s'" % client_name)
+        canonicalizer = canonicalizers[client_name]
+        full_trace = readOutput(proc_info)
+        test.storeTrace(client_name, full_trace,proc_info['cmd'])
+        canon_trace = [VMUtils.toText(step) for step in canonicalizer(full_trace.split("\n"))]
+        test.canon_traces.append(canon_trace)
+        tracelen = len(canon_trace)
+        logger.info("Processed %s steps for %s on test %s" % (tracelen, client_name, test.identifier))
+        
+    return tracelen
 
 def processTraces(test):
     if test is None:
-        return True
+        return None
 
     # Process previous traces
 
     (equivalent, trace_output) = VMUtils.compare_traces(test.canon_traces, cfg.clientNames()) 
 
-    if equivalent and False:
+    if equivalent:
         test.removeFiles()
+        return None
     else:
         logger.warning("CONSENSUS BUG!!!")
         trace_summary = get_summary(trace_output)
@@ -395,65 +397,92 @@ def processTraces(test):
         test.addArtefact("combined_trace.log","\n".join(trace_output))
         test.addArtefact("shortened_trace.log","\n".join(trace_summary))
 
-    return equivalent
+    return test
 
-def perform_tests(test_iterator):
-    
-    pass_count = 0
-    fail_count = 0
-    failures = []
+class TestExecutor():
 
-    previous_test = None
+    def __init__(self):
+        self.pass_count = 0
+        self.fail_count = 0
+        self.failures = []
+        self.traceLengths = collections.deque([],100)
+        self.start_time = time.time()
 
-    start_time = time.time()
 
-    n = 0
+    def startFuzzing(self):   
+        previous_test = None
+        self.start_time = time.time()
+        n = 0
 
-    def __end_previous_test():
-        nonlocal n, fail_count, pass_count
-        global traceFiles
+        def __end_previous_test():
+            nonlocal n
 
-        # End previous procs
-        traceFiles = end_processes(previous_test)
+            # End previous procs
+            traceLength = end_processes(previous_test)
+            if traceLength is not None:
+                self.traceLengths.append(traceLength)
 
-        # Process previous traces
-        if processTraces(previous_test):
-            pass_count = pass_count +1
-        else:
-            fail_count = fail_count +1
+            # Process previous traces
+            failingTestcase = processTraces(previous_test)
+            if failingTestcase is None:
+                self.pass_count = self.pass_count +1
+            else:
+                self.fail_count = self.fail_count +1
+                self.failures.append(failingTestcase.listArtefacts())
+ 
+            # Do some reporting
+            if n % 10 == 0:
+                time_elapsed = time.time() - self.start_time
+                logger.info("Fails: {}, Pass: {}, #test {} speed: {:f} tests/s".format(
+                        self.fail_count, 
+                        self.pass_count, 
+                        (self.fail_count + self.pass_count),
+                        (self.fail_count + self.pass_count) / time_elapsed
+                    ))
 
-        # Do some reporting
+        for test in generateTests():
+            n = n+1
+            #Prepare the current test
+            logger.info("Test id: %s" % test.id())
+            test.writeToFile()
 
-        if n % 10 == 0:
-            time_elapsed = time.time() - start_time
-            logger.info("Fails: {}, Pass: {}, #test {} speed: {:f} tests/s".format(
-                    fail_count, 
-                    pass_count, 
-                    (fail_count + pass_count),
-                    (fail_count + pass_count) / time_elapsed
-                ))
+            # Start new procs
+            start_processes(test)
 
-    for test in test_iterator():
-        n = n+1
-        #Prepare the current test
-        logger.info("Test id: %s" % test.id())
-        test.writeToFile()
+            __end_previous_test()
 
-        # Start new procs
-        start_processes(test)
+            previous_test = test
 
         __end_previous_test()
 
-        previous_test = test
+        return (n, len(failures), pass_count, failures)
 
-    __end_previous_test()
-
-    return (n, len(failures), pass_count, failures)
+    def status(self):
+        import collections, statistics
+        from datetime import datetime
+        return {
+            "starttime" : datetime.utcfromtimestamp(self.start_time).strftime('%Y-%m-%d %H:%M:%S'),
+            "pass" : self.pass_count, 
+            "fail" : self.fail_count,
+            "failures" : self.failures,
+            "speed" : (self.fail_count + self.pass_count) / (time.time() - self.start_time),
+            "mean" : statistics.mean(self.traceLengths) if self.traceLengths else "NA",
+            "stdev" : statistics.stdev(self.traceLengths) if len(self.traceLengths) >2 else "NA",
+            "numZero" : self.traceLengths.count(0) if self.traceLengths else "NA" ,
+            "max" : max(self.traceLengths) if self.traceLengths else "NA",
+        }
 
 def testSummary():
     """Enable this, and test by passing a trace-output via console"""
     with open(sys.argv[1]) as f:
         print("".join(get_summary(f.readlines())))
+
+def main():
+    # Start all docker daemons that we'll use during the execution
+    startDaemons()
+
+    TestExecutor().startFuzzing()
+
 
 if __name__ == '__main__':
 #    testSummary()
